# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Rest Clon de Airbnb",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/users": {
            "get": {
                "description": "Get all users",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "We get all users succesfully"
                    },
                    "400": {
                        "description": "Generate an error "
                    }
                }
            }
        },
...
        "api/v1/accommodations/": {
            "get": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "type": "string"
                    }
                ],
                "description": "Obtener al usuario loggeado",
                "tags": [
                    "accomodation"
                ],
                "responses": {
                    "200": {
                        "description": "Obtener exitosamente el usuario"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la peticion no se resuelva"
                    }
                }
            }
        },
...
        "api/v1/reservations/": {
            "get": {
                "description": "Obtener las reservaciones del usuario",
                "tags": [
                    "reservation"
                ],
                "responses": {
                    "200": {
                        "description": "Obtener exitosamente las reservaciones del usuario"
                    },
                    "400": {
                        "description": "Generamos un error en caso de que la peticion no se resuelva"
                    }
                }
            }
        },
...
    }
}